refactor(#3282): extract __any_eq/__any_strict_eq + relational family to any-eq-helpers.ts (slice B)
Lift the `any`-typed equality & relational-comparison family
(`__any_eq`, `__any_strict_eq`, `__any_lt`/`__any_gt`/`__any_le`/`__any_ge`)
verbatim out of the `ensureAnyHelpers` god-function into a new sibling module
`src/codegen/any-eq-helpers.ts` via `registerAnyEqHelpers`, mirroring slice A's
`registerAnyBoxHelpers` pattern.

`addHelper` is threaded in as a callback and the tag-5 coercion/equality
closures (`tag5ToNumber` / `tag5ValueEqThen`) are threaded as params so their
captured environment is unchanged. The standalone/wasi-gated #2175 V2-S3
reference-identity reconciliation inside `__any_strict_eq` moves verbatim WITH
its `ctx.standalone || ctx.wasi` guard, and registration order
(eq -> strict_eq -> lt -> gt -> le -> ge) is preserved exactly, so funcIdx
assignment — and therefore the emitted Wasm — is unchanged.

- prove-emit-identity check: IDENTICAL 60/60 across gc/standalone/wasi/linear.
- any-helpers.ts drops 2425 -> 1926 LOC (-499); no import back into it (the eq
  region references nothing from any-helpers.ts) so there is no ESM cycle.
- tsc --noEmit 0; biome lint clean; check:godfiles OK (none new/grown).

Epic #3282 remains open for the arithmetic and `__any_typeof` tail families.

✓
